Tile decoders refactoring
