+++ vtm-mvt/src/org/oscim/tiling/source/mvt/OpenMapTilesMvtTileSource.java
@@ -69,6 +69,6 @@ public class OpenMapTilesMvtTileSource extends UrlTileSource {
 
     @Override
     public ITileDataSource getDataSource() {
-        return new OverzoomTileDataSource(new UrlTileDataSource(this, new MvtTileDecoder(locale), getHttpEngine()), mOverZoom);
+        return new OverzoomTileDataSource(new UrlTileDataSource(this, new TileDecoder(locale), getHttpEngine()), mOverZoom);
     }
 }
