+++ vtm/src/org/oscim/tiling/source/mapzen/TileDecoder.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License along with
  * this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.oscim.tiling.source.mvt;
+package org.oscim.tiling.source.mapzen;
 
 import org.oscim.core.GeometryBuffer.GeometryType;
 import org.oscim.core.MapElement;
